OpenMPTarget: Moved the Kokkos_OpenMPTarget.hpp file inside OpenMPTarget dir.
