+++ core/src/OpenMPTarget/Kokkos_OpenMPTargetSpace.cpp
@@ -33,7 +33,7 @@
 #include <sstream>
 #include <cstring>
 
-#include <Kokkos_OpenMPTarget.hpp>
+#include <OpenMPTarget/Kokkos_OpenMPTarget.hpp>
 #include <OpenMPTarget/Kokkos_OpenMPTargetSpace.hpp>
 #include <impl/Kokkos_Error.hpp>
 #include <Kokkos_Atomic.hpp>
